+++ src/Listener.php
@@ -7,11 +7,13 @@
  * Company: Pronamic
  *
  * @author Remco Tolsma
- * @version 1.0.1
+ * @version 1.1.4
+ * @since 1.0.1
  */
 class Pronamic_WP_Pay_Gateways_IDealBasic_Listener implements Pronamic_Pay_Gateways_ListenerInterface {
 	public static function listen() {
-		if ( filter_has_var( INPUT_GET, 'xml_notifaction' ) ) {
+		// Also check for typo 'xml_notifaction', as this has been used in the past.
+		if ( filter_has_var( INPUT_GET, 'xml_notification' ) || filter_has_var( INPUT_GET, 'xml_notifaction' ) ) {
 			$data = file_get_contents( 'php://input' );
 
 			$xml = Pronamic_WP_Util::simplexml_load_string( $data );