+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## [Unreleased][unreleased]
 - 
 
+## [1.1.4] - 2016-03-22
+- Updated gateway settings.
+- Fixed typo 'xml_notifaction' in listener parameter.
+
 ## [1.1.3] - 2016-03-02
 - Added an get_settings function to the integration class.
 - Moved get_gateway_class() function to the configuration class.
@@ -32,3 +36,12 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 
 ## 1.0.0 - 2015-01-19
 - First release.
+
+[unreleased]: https://github.com/wp-pay-gateways/ideal-basic/compare/1.1.4...HEAD
+[1.1.4]: https://github.com/wp-pay-gateways/ideal-basic/compare/1.1.3...1.1.4
+[1.1.3]: https://github.com/wp-pay-gateways/ideal-basic/compare/1.1.2...1.1.3
+[1.1.2]: https://github.com/wp-pay-gateways/ideal-basic/compare/1.1.1...1.1.2
+[1.1.1]: https://github.com/wp-pay-gateways/ideal-basic/compare/1.1.0...1.1.1
+[1.1.0]: https://github.com/wp-pay-gateways/ideal-basic/compare/1.0.2...1.1.0
+[1.0.2]: https://github.com/wp-pay-gateways/ideal-basic/compare/1.0.1...1.0.2
+[1.0.1]: https://github.com/wp-pay-gateways/ideal-basic/compare/1.0.0...1.0.1